% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_SAS_procedure.R
\name{get_SAS_procedure}
\alias{get_SAS_procedure}
\title{get SAS procedures}
\usage{
get_SAS_procedure(text)
}
\arguments{
\item{text}{SAS code}
}
\value{
list of identified procedures as
-proc : name of the procedure in listed values in procs
-contenu : what's inside the procedure
}
\description{
read SAS code, identify procedures and extract their contents
}
\examples{
get_SAS_procedure(text = "PROC CONTENTS DATA=sample;RUN;")
}
