% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ProcMeans.R
\docType{data}
\name{proc_means}
\alias{proc_means}
\title{proc_means}
\format{
An object of class \code{character} of length 1.
}
\usage{
proc_means
}
\arguments{
\item{code_sas}{}
}
\value{
code R
}
\description{
proc_means
}
\keyword{datasets}
