% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ProcContents.R
\name{ProcContents}
\alias{ProcContents}
\title{Parsed Representation of a SAS PROC CONTENTS Statement}
\usage{
ProcContents(code_sas)
}
\description{
`ProcContents` is an S7 class representing a parsed SAS `PROC CONTENTS`
statement. It extends [ProcSAS] and provides a concrete implementation
of the [transpile()] method.
}
\details{
This class inherits all parsing behavior from [ProcSAS]. The `transpile()`
method translates a `PROC CONTENTS` statement into equivalent R code,
returning a character string representing the generated R expression.
}
\section{Inheritance}{

`ProcContents` inherits from [ProcSAS] and therefore exposes the same
properties:
\describe{
  \item{source}{Original SAS code supplied to the constructor.}
  \item{proc_data}{Name of the dataset specified in the `DATA=` argument.}
  \item{proc_options}{Character vector of remaining procedure options.}
}
}

\section{Methods}{

\describe{
  \item{transpile()}{Translate the SAS `PROC CONTENTS` statement into R code.}
}
}

\examples{
sas_code <- "proc contents data=mydata;"
proc <- ProcContents(sas_code)
proc

}
\seealso{
[ProcSAS] for the parent class,
[transpile()] for the generic interface.

Other SAS procedure classes: 
\code{\link{ProcMeans}()},
\code{\link{ProcPrint}()},
\code{\link{ProcSAS}()}
}
\concept{SAS procedure classes}
\keyword{internal}
