% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ProcFreq.R
\name{ProcFreq}
\alias{ProcFreq}
\title{Parsed Representation of a SAS PROC FREQ Statement}
\usage{
ProcFreq(sas_code)
}
\description{
`ProcFreq` is an S7 class representing a parsed SAS `PROC FREQ`
statement. It extends [ProcSAS] and implements procedure-specific
parsing and transpilation behavior.
}
\details{
In addition to the core metadata extracted by [ProcSAS], this class
parses common `PROC FREQ` statements such as `VAR`, `BY`, `CLASS`,
`WEIGHT`, and `OUTPUT`. These components are stored as structured
properties and can be used by the `transpile()` method or downstream
workflows.

The `transpile()` method translates a `PROC FREQ` statement into
equivalent R code. The returned value is a character string representing
the generated R expression.
}
\section{Inheritance}{

`ProcFreq` inherits from [ProcSAS] and therefore exposes the same
base properties:
\describe{
  \item{source}{Original SAS code supplied to the constructor.}
  \item{proc_data}{Name of the dataset specified in the `DATA=` argument.}
  \item{proc_options}{Character vector of remaining procedure options.}
}
}

\section{Additional Properties}{

\describe{
  \item{pf_by}{Variables specified in the `BY` statement.}
  \item{pf_exact}{Frequency variable specified in the `EXACT` statement.}
  \item{pf_tables}{Frequency variable specified in the `TABLES` statement.}
  \item{pf_test}{Frequency variable specified in the `TEST` statement.}
  \item{pf_weight}{Weight variable specified in the `WEIGHT` statement.}
  \item{pf_output}{List describing the `OUTPUT` statement.}
}
}

\section{Methods}{

\describe{
  \item{transpile()}{Translate the SAS `PROC FREQ` statement into R code.}
}
}

\examples{
sas_code <- "
proc FREQ data=mydata;
  tables y * x;
  output out=table1;
run;
"
proc <- ProcFreq(sas_code)
proc

}
\seealso{
[ProcSAS] for the parent class,
[transpile()] for the generic interface.

Other SAS procedure classes: 
\code{\link{ProcContents}()},
\code{\link{ProcMeans}()},
\code{\link{ProcPrint}()},
\code{\link{ProcSAS}()},
\code{\link{ProcSQL}()},
\code{\link{ProcSort}()}
}
\concept{SAS procedure classes}
\keyword{internal}
