% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ProcSort.R
\name{ProcSort}
\alias{ProcSort}
\title{Parsed Representation of a SAS PROC SORT Statement}
\usage{
ProcSort(code_sas)
}
\description{
`ProcSort` is an S7 class representing a parsed SAS `PROC SORT`
statement. It extends [ProcSAS] and implements procedure-specific
parsing and transpilation behavior.
}
\details{
In addition to the core metadata extracted by [ProcSAS], this class
parses common `PROC SORT` statements such as `BY`.
These components are stored as structured
properties and can be used by the `transpile()` method or downstream
workflows.

The `transpile()` method translates a `PROC SORT` statement into
equivalent R code. The returned value is a character string representing
the generated R expression.
}
\section{Inheritance}{

`ProcSort` inherits from [ProcSAS] and therefore exposes the same
base properties:
\describe{
  \item{source}{Original SAS code supplied to the constructor.}
  \item{proc_data}{Name of the dataset specified in the `DATA=` argument.}
  \item{proc_options}{Character vector of remaining procedure options.}
}
}

\section{Additional Properties}{

\describe{
  \item{pm_by}{Variables specified in the `BY` statement.}
}
}

\section{Methods}{

\describe{
  \item{transpile()}{Translate the SAS `PROC SORT` statement into R code.}
}
}

\examples{
sas_code <- "
proc sort data = iris;
by  descending Sepal.Lenght Petal.Lenght;
run;"
proc <- ProcSort(sas_code)
proc

}
\seealso{
[ProcSAS] for the parent class,
[transpile()] for the generic interface.

Other SAS procedure classes: 
\code{\link{ProcContents}()},
\code{\link{ProcMeans}()},
\code{\link{ProcPrint}()},
\code{\link{ProcSAS}()}
}
\concept{SAS procedure classes}
\keyword{internal}
