% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_SAS_proc_means.R
\name{read_SAS_proc_means}
\alias{read_SAS_proc_means}
\title{read PROC MEANS}
\usage{
read_SAS_proc_means(text)
}
\arguments{
\item{text}{}
}
\description{
The means procedure is written as follow :
PROC MEANS <options> <statistic-keyword(s)>;
BY <DESCENDING> variable-1 <<DESCENDING> variable-2 ...> <NOTSORTED>;
CLASS variable(s) </ options>;
FREQ variable;
ID variable(s);
OUTPUT <OUT=SAS-data-set> <output-statistic-specification(s)>
  <id-group-specification(s)> <maximum-id-specification(s)>
  <minimum-id-specification(s)> </ options> ;
TYPES request(s);
VAR variable(s) </ WEIGHT=weight-variable>;
WAYS list;
WEIGHT variable;
source : https://documentation.sas.com/doc/en/pgmsascdc/9.4_3.5/proc/p0f0fjpjeuco4gn1ri963f683mi4.htm

This function read a cut sas means procedure, identify each parameter with regex and returns their values.
ID, OUTPUT, TYPES and WAYS statements are irrevelant to translate to R. Thus they are not read.
}
