% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_SAS_procedures.R
\name{split_SAS_procedures}
\alias{split_SAS_procedures}
\title{split_SAS_procedures}
\usage{
split_SAS_procedures(text)
}
\arguments{
\item{text}{character: code sas}
}
\value{
character vector, procedures splitted
}
\description{
split_SAS_procedures
}
