% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ProcSQL.R
\name{sql_to_dplyr}
\alias{sql_to_dplyr}
\title{sql_to_dplyr}
\usage{
sql_to_dplyr(sql_code)
}
\arguments{
\item{sql_code}{: chaine de charactère code SQL}
}
\value{
chaine de charactere
}
\description{
sql_to_dplyr
}
