% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transastor.R
\name{transastor}
\alias{transastor}
\title{Translate SAS code to R}
\usage{
transastor(sas_code)
}
\arguments{
\item{sas_code}{Character string containing SAS code.}
}
\value{
A character string containing translated R code.
}
\description{
Functions used to translate (transpile) SAS code into R code.
The main entry point is \code{transastor()}, which parses SAS code,
dispatches each SAS procedure to the appropriate translator,
and replaces the original code with its R equivalent.
}
\examples{
sas_code <- "
proc means data=mydata;
run;
"

transastor(sas_code)

}
