% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ProcSQL.R
\name{ProcSQL}
\alias{ProcSQL}
\title{Parsed Representation of a SAS PROC SQL Statement}
\usage{
ProcSQL(sql_code)
}
\description{
`ProcSQL` is an S7 class representing a parsed SAS `PROC SQL`
statement. It extends [ProcSAS] and implements procedure-specific
transpilation behavior.
}
\details{
The `transpile()` method translates a `PROC SQL` statement into
equivalent R code. The returned value is a character string representing
the generated R expression.
}
\section{Inheritance}{

`ProcSQL` inherits from [ProcSAS] and therefore exposes the same
properties:
\describe{
  \item{source}{Original SAS code supplied to the constructor.}
  \item{proc_data}{Name of the dataset specified in the `DATA=` argument.}
  \item{proc_options}{Character vector of remaining procedure options.}
}
}

\section{Methods}{

\describe{
  \item{transpile()}{Translate the SAS `PROC SQL` statement into R code.}
}
}

\examples{
sql_code <- "PROC SQL; select var1, var2, var3 from table; QUIT;"
proc <- ProcSQL(sql_code)
proc

}
\seealso{
[ProcSAS] for the parent class,
[transpile()] for the generic interface.

Other SAS procedure classes: 
\code{\link{ProcContents}()},
\code{\link{ProcFreq}()},
\code{\link{ProcMeans}()},
\code{\link{ProcPrint}()},
\code{\link{ProcSAS}()},
\code{\link{ProcSort}()}
}
\concept{SAS procedure classes}
\keyword{internal}
